+++ UnionArgParser/assemblyInfo.fs
@@ -1,9 +1,9 @@
 ﻿namespace System
 open System.Reflection
 
-[<assembly: AssemblyVersionAttribute("0.5.9")>]
-[<assembly: AssemblyFileVersionAttribute("0.5.9")>]
+[<assembly: AssemblyVersionAttribute("0.6.0")>]
+[<assembly: AssemblyFileVersionAttribute("0.6.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.5.9"
+    let [<Literal>] Version = "0.6.0"