+++ RELEASE_NOTES.md
@@ -1,2 +1,11 @@
+#### 0.6.0
+* Update root namespace to Nessos.UnionArgParser
+* Add support for application configuration in class libraries
+* Change application configuration parsing API.
+
 #### 0.5.9
-* Add FAKE support.
+* Fix packaging issue.
+
+#### 0.5.8
+* Include optional BindingFlags parameter. 
+* Include support for all primitive types.